fix typo with options, remove logs from tests
