+++ lib/methods/export.js
@@ -21,7 +21,7 @@ module.exports = self => {
       const ids = self.apos.launder.ids(req.body._ids);
       const relatedTypes = self.apos.launder.strings(req.body.relatedTypes);
       const expiration = self.options.importExport?.export?.expiration &&
-        self.apos.launder.integer(self.options.export.expiration);
+        self.apos.launder.integer(self.options.importExport.export.expiration);
 
       const [ defaultFormatName ] = Object.keys(self.formats);
       const formatName = self.apos.launder.string(req.body.formatName, defaultFormatName);
