+++ test/index.js
@@ -30,7 +30,18 @@ describe('@apostrophecms/import-export', function () {
     apos = await t.create({
       root: module,
       testModule: true,
-      modules: getAppConfig()
+      modules: {
+        ...getAppConfig(),
+        '@apostrophecms/import-export': {
+          options: {
+            importExport: {
+              export: {
+                expiration: 10 * 1000
+              }
+            }
+          }
+        }
+      }
     });
 
     tempPath = path.join(apos.rootDir, 'data/temp/uploadfs');
@@ -38,7 +49,6 @@ describe('@apostrophecms/import-export', function () {
     exportsPath = path.join(apos.rootDir, 'public/uploads/exports');
     importExportManager = apos.modules['@apostrophecms/import-export'];
     importExportManager.removeFromUploadFs = () => {};
-    importExportManager.remove = () => {};
     gzip = importExportManager.formats.gzip;
     mimeType = gzip.allowedTypes[0];
 
@@ -56,7 +66,7 @@ describe('@apostrophecms/import-export', function () {
   afterEach(async function() {
     await deletePiecesAndPages(apos);
     await deleteAttachments(apos, attachmentPath);
-    await cleanData([ tempPath, exportsPath, attachmentPath ]);
+    await cleanData([ tempPath, exportsPath ]);
   });
 
   it('should generate a zip file for pieces without related documents', async function () {
@@ -263,8 +273,9 @@ describe('@apostrophecms/import-export', function () {
 
     const { url } = await importExportManager.export(req, manager);
     const fileName = path.basename(url);
-
-    piecesTgzPath = path.join(exportsPath, fileName);
+    const exportFilePath = path.join(exportsPath, fileName);
+    const importFilePath = path.join(tempPath, fileName);
+    await fs.copyFile(exportFilePath, importFilePath);
 
     await deletePiecesAndPages(apos);
     await deleteAttachments(apos, attachmentPath);
@@ -272,7 +283,7 @@ describe('@apostrophecms/import-export', function () {
     req.body = {};
     req.files = {
       file: {
-        path: piecesTgzPath,
+        path: importFilePath,
         type: mimeType
       }
     };
@@ -349,13 +360,14 @@ describe('@apostrophecms/import-export', function () {
 
     const { url } = await importExportManager.export(req, manager);
     const fileName = path.basename(url);
-
-    piecesTgzPath = path.join(exportsPath, fileName);
+    const exportFilePath = path.join(exportsPath, fileName);
+    const importFilePath = path.join(tempPath, fileName);
+    await fs.copyFile(exportFilePath, importFilePath);
 
     req.body = {};
     req.files = {
       file: {
-        path: piecesTgzPath,
+        path: importFilePath,
         type: mimeType
       }
     };
@@ -460,8 +472,9 @@ describe('@apostrophecms/import-export', function () {
 
     const { url } = await importExportManager.export(req, apos.page);
     const fileName = path.basename(url);
-
-    pageTgzPath = path.join(exportsPath, fileName);
+    const exportFilePath = path.join(exportsPath, fileName);
+    const importFilePath = path.join(tempPath, fileName);
+    await fs.copyFile(exportFilePath, importFilePath);
 
     await deletePiecesAndPages(apos);
     await deleteAttachments(apos, attachmentPath);
@@ -469,7 +482,7 @@ describe('@apostrophecms/import-export', function () {
     req.body = {};
     req.files = {
       file: {
-        path: pageTgzPath,
+        path: importFilePath,
         type: mimeType
       }
     };
@@ -513,13 +526,14 @@ describe('@apostrophecms/import-export', function () {
 
     const { url } = await importExportManager.export(req, apos.page);
     const fileName = path.basename(url);
-
-    pageTgzPath = path.join(exportsPath, fileName);
+    const exportFilePath = path.join(exportsPath, fileName);
+    const importFilePath = path.join(tempPath, fileName);
+    await fs.copyFile(exportFilePath, importFilePath);
 
     req.body = {};
     req.files = {
       file: {
-        path: pageTgzPath,
+        path: importFilePath,
         type: mimeType
       }
     };
@@ -618,13 +632,14 @@ describe('@apostrophecms/import-export', function () {
 
     const { url } = await importExportManager.export(req, apos.page);
     const fileName = path.basename(url);
-
-    pageTgzPath = path.join(exportsPath, fileName);
+    const exportFilePath = path.join(exportsPath, fileName);
+    const importFilePath = path.join(tempPath, fileName);
+    await fs.copyFile(exportFilePath, importFilePath);
 
     req.body = {};
     req.files = {
       file: {
-        path: pageTgzPath,
+        path: importFilePath,
         type: mimeType
       }
     };
@@ -746,8 +761,9 @@ describe('@apostrophecms/import-export', function () {
 
     const { url } = await importExportManager.export(req, apos.page);
     const fileName = path.basename(url);
-
-    pageTgzPath = path.join(exportsPath, fileName);
+    const exportFilePath = path.join(exportsPath, fileName);
+    const importFilePath = path.join(tempPath, fileName);
+    await fs.copyFile(exportFilePath, importFilePath);
 
     // Now that it's exported as draft, PUBLISH the page again
     const { lastPublishedAt } = await apos.page.publish(req, draftPage);
@@ -756,7 +772,7 @@ describe('@apostrophecms/import-export', function () {
     req.body = {};
     req.files = {
       file: {
-        path: pageTgzPath,
+        path: importFilePath,
         type: mimeType
       }
     };
@@ -898,595 +914,4 @@ describe('@apostrophecms/import-export', function () {
     });
   });
 
-  describe('#import - man-made CSV file', function() {
-    let req;
-    let notify;
-    let input;
-    let csv;
-
-    this.beforeEach(async function() {
-      csv = importExportManager.formats.csv;
-      mimeType = csv.allowedTypes[0];
-
-      req = apos.task.getReq({
-        locale: 'en',
-        body: {},
-        files: {
-          file: {
-            path: '/some/path/to/file',
-            type: mimeType
-          }
-        }
-      });
-      notify = apos.notify;
-      input = csv.input;
-
-      await deletePiecesAndPages(apos);
-      await deleteAttachments(apos, attachmentPath);
-    });
-
-    this.afterEach(function() {
-      apos.notify = notify;
-      csv.input = input;
-    });
-
-    it('should notify when the type is not provided', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              title: 'topic1',
-              description: 'description1',
-              main: '<p><em>rich</em> <strong>text</strong></p>'
-            }
-          ]
-        };
-      };
-
-      const messages = [];
-
-      apos.notify = async (req, message, options) => {
-        messages.push(message);
-        return notify(req, message, options);
-      };
-
-      await importExportManager.import(req);
-
-      assert.equal(messages.some(message => message === 'aposImportExport:typeUnknown'), true);
-    });
-
-    it('should notify when the type does not exist', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'random-type',
-              title: 'topic1',
-              description: 'description1',
-              main: '<p><em>rich</em> <strong>text</strong></p>'
-            }
-          ]
-        };
-      };
-
-      const messages = [];
-
-      apos.notify = async (req, message, options) => {
-        messages.push(message);
-        return notify(req, message, options);
-      };
-
-      await importExportManager.import(req);
-
-      assert.equal(messages.some(message => message === 'aposImportExport:typeUnknown'), true);
-    });
-
-    it('should import a piece from a csv file that was not made from the import-export module', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'topic',
-              title: 'topic1',
-              description: 'description1',
-              main: '<p><em>rich</em> <strong>text</strong></p>'
-            }
-          ]
-        };
-      };
-
-      await importExportManager.import(req);
-
-      const topics = await apos.doc.db
-        .find({ type: 'topic' })
-        .toArray();
-
-      assert.equal(topics.length, 1);
-      assert.equal(topics[0].title, 'topic1');
-      assert.equal(topics[0].slug, 'topic1');
-      assert.equal(topics[0].aposMode, 'draft');
-      assert.equal(topics[0].description, 'description1');
-      assert.equal(topics[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong></p>');
-    });
-
-    it('should import a page from a csv file that was not made from the import-export module', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'default-page',
-              title: 'page1',
-              main: '<p><em>rich</em> <strong>text</strong></p>'
-            }
-          ]
-        };
-      };
-
-      await importExportManager.import(req);
-
-      const pages = await apos.doc.db
-        .find({ type: 'default-page' })
-        .toArray();
-
-      assert.equal(pages.length, 1);
-      assert.equal(pages[0].title, 'page1');
-      assert.equal(pages[0].slug, '/page1');
-      assert.equal(pages[0].aposMode, 'draft');
-      assert.equal(pages[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong></p>');
-    });
-
-    it('should insert a piece as draft and published when there is an update key that does not match any existing doc', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'topic',
-              'title:key': 'topic1',
-              title: 'topic1 - edited',
-              description: 'description1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      await importExportManager.import(req);
-
-      const topics = await apos.doc.db
-        .find({ type: 'topic' })
-        .toArray();
-
-      assert.equal(topics.length, 2);
-
-      assert.equal(topics[0].title, 'topic1 - edited');
-      assert.equal(topics[0].slug, 'topic1-edited');
-      assert.equal(topics[0].aposMode, 'draft');
-      assert.equal(topics[0].description, 'description1 - edited');
-      assert.equal(topics[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-
-      assert.equal(topics[1].title, 'topic1 - edited');
-      assert.equal(topics[1].slug, 'topic1-edited');
-      assert.equal(topics[1].aposMode, 'published');
-      assert.equal(topics[1].description, 'description1 - edited');
-      assert.equal(topics[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-
-    it('should insert a page as draft and published when there is an update key that does not match any existing doc', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'default-page',
-              'title:key': 'page1',
-              title: 'page1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      await importExportManager.import(req);
-
-      const pages = await apos.doc.db
-        .find({ type: 'default-page' })
-        .toArray();
-
-      assert.equal(pages.length, 2);
-
-      assert.equal(pages[0].title, 'page1 - edited');
-      assert.equal(pages[0].slug, '/page1-edited');
-      assert.equal(pages[0].aposMode, 'draft');
-      assert.equal(pages[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-
-      assert.equal(pages[1].title, 'page1 - edited');
-      assert.equal(pages[1].slug, '/page1-edited');
-      assert.equal(pages[1].aposMode, 'published');
-      assert.equal(pages[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-
-    it('should insert a piece as draft and published when there is an empty update key', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'topic',
-              'title:key': '',
-              title: 'topic1 - edited',
-              description: 'description1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      await importExportManager.import(req);
-
-      const topics = await apos.doc.db
-        .find({ type: 'topic' })
-        .toArray();
-
-      assert.equal(topics.length, 2);
-
-      assert.equal(topics[0].title, 'topic1 - edited');
-      assert.equal(topics[0].slug, 'topic1-edited');
-      assert.equal(topics[0].aposMode, 'draft');
-      assert.equal(topics[0].description, 'description1 - edited');
-      assert.equal(topics[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-
-      assert.equal(topics[1].title, 'topic1 - edited');
-      assert.equal(topics[1].slug, 'topic1-edited');
-      assert.equal(topics[1].aposMode, 'published');
-      assert.equal(topics[1].description, 'description1 - edited');
-      assert.equal(topics[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-
-    it('should insert a page as draft and published when there is an empty update key', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'default-page',
-              'title:key': '',
-              title: 'page1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      await importExportManager.import(req);
-
-      const pages = await apos.doc.db
-        .find({ type: 'default-page' })
-        .toArray();
-
-      assert.equal(pages.length, 2);
-
-      assert.equal(pages[0].title, 'page1 - edited');
-      assert.equal(pages[0].slug, '/page1-edited');
-      assert.equal(pages[0].aposMode, 'draft');
-      assert.equal(pages[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-
-      assert.equal(pages[1].title, 'page1 - edited');
-      assert.equal(pages[1].slug, '/page1-edited');
-      assert.equal(pages[1].aposMode, 'published');
-      assert.equal(pages[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-
-    it('should update a piece draft and published versions when there is an update key that matches an existing doc', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'topic',
-              'title:key': 'topic1',
-              title: 'topic1 - edited',
-              description: 'description1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      const topic = await apos.topic.insert(req, {
-        ...apos.topic.newInstance(),
-        title: 'topic1',
-        description: 'description1',
-        main: '<p><em>rich</em> <strong>text</strong></p>'
-      });
-
-      await importExportManager.import(req);
-
-      const topics = await apos.doc.db
-        .find({ type: 'topic' })
-        .toArray();
-
-      assert.equal(topics.length, 2);
-
-      assert.equal(topics[0].aposDocId, topic.aposDocId);
-      assert.equal(topics[0].title, 'topic1 - edited');
-      assert.equal(topics[0].slug, 'topic1');
-      assert.equal(topics[0].aposMode, 'draft');
-      assert.equal(topics[0].description, 'description1 - edited');
-      assert.equal(topics[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-      assert.equal(topics[0].modified, false);
-
-      assert.equal(topics[1].aposDocId, topic.aposDocId);
-      assert.equal(topics[1].title, 'topic1 - edited');
-      assert.equal(topics[1].slug, 'topic1');
-      assert.equal(topics[1].aposMode, 'published');
-      assert.equal(topics[1].description, 'description1 - edited');
-      assert.equal(topics[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-
-    it('should update a page draft and published versions when there is an update key that matches an existing doc', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'default-page',
-              'title:key': 'page1',
-              title: 'page1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      const page = await apos.page.insert(req, '_home', 'lastChild', {
-        ...apos.modules['default-page'].newInstance(),
-        title: 'page1',
-        main: '<p><em>rich</em> <strong>text</strong></p>'
-      });
-
-      await importExportManager.import(req);
-
-      const pages = await apos.doc.db
-        .find({ type: 'default-page' })
-        .toArray();
-
-      assert.equal(pages.length, 2);
-
-      assert.equal(pages[0].aposDocId, page.aposDocId);
-      assert.equal(pages[0].title, 'page1 - edited');
-      assert.equal(pages[0].slug, '/page1');
-      assert.equal(pages[0].aposMode, 'draft');
-      assert.equal(pages[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-      assert.equal(pages[0].modified, false);
-
-      assert.equal(pages[1].aposDocId, page.aposDocId);
-      assert.equal(pages[1].title, 'page1 - edited');
-      assert.equal(pages[1].slug, '/page1');
-      assert.equal(pages[1].aposMode, 'published');
-      assert.equal(pages[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-
-    it('should update a piece draft and published versions when there is an update key that only matches the existing draft doc', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'topic',
-              'title:key': 'topic1',
-              title: 'topic1 - edited',
-              description: 'description1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      const topic = await apos.topic.insert(req, {
-        ...apos.topic.newInstance(),
-        title: 'topic1',
-        description: 'description1',
-        main: '<p><em>rich</em> <strong>text</strong></p>'
-      });
-
-      // check that the published doc is also updated, even with a different title
-      await apos.doc.db.updateOne(
-        {
-          aposDocId: topic.aposDocId,
-          aposMode: 'published'
-        },
-        {
-          $set: {
-            title: 'topic1 - published title that does not match the draft title nor the update key'
-          }
-        }
-      );
-
-      await importExportManager.import(req);
-
-      const topics = await apos.doc.db
-        .find({ type: 'topic' })
-        .toArray();
-
-      assert.equal(topics.length, 2);
-
-      assert.equal(topics[0].aposDocId, topic.aposDocId);
-      assert.equal(topics[0].title, 'topic1 - edited');
-      assert.equal(topics[0].slug, 'topic1');
-      assert.equal(topics[0].aposMode, 'draft');
-      assert.equal(topics[0].description, 'description1 - edited');
-      assert.equal(topics[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-      assert.equal(topics[0].modified, false);
-
-      assert.equal(topics[1].aposDocId, topic.aposDocId);
-      assert.equal(topics[1].title, 'topic1 - edited');
-      assert.equal(topics[1].slug, 'topic1');
-      assert.equal(topics[1].aposMode, 'published');
-      assert.equal(topics[1].description, 'description1 - edited');
-      assert.equal(topics[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-
-    it('should update a page draft and published versions when there is an update key that only matches the existing draft doc', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'default-page',
-              'title:key': 'page1',
-              title: 'page1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      const page = await apos.page.insert(req, '_home', 'lastChild', {
-        ...apos.modules['default-page'].newInstance(),
-        title: 'page1',
-        main: '<p><em>rich</em> <strong>text</strong></p>'
-      });
-
-      // check that the published doc is also updated, even with a different title
-      await apos.doc.db.updateOne(
-        {
-          aposDocId: page.aposDocId,
-          aposMode: 'published'
-        },
-        {
-          $set: {
-            title: 'page1 - published title that does not match the draft title nor the update key'
-          }
-        }
-      );
-
-      await importExportManager.import(req);
-
-      const pages = await apos.doc.db
-        .find({ type: 'default-page' })
-        .toArray();
-
-      assert.equal(pages.length, 2);
-
-      assert.equal(pages[0].aposDocId, page.aposDocId);
-      assert.equal(pages[0].title, 'page1 - edited');
-      assert.equal(pages[0].slug, '/page1');
-      assert.equal(pages[0].aposMode, 'draft');
-      assert.equal(pages[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-      assert.equal(pages[0].modified, false);
-
-      assert.equal(pages[1].aposDocId, page.aposDocId);
-      assert.equal(pages[1].title, 'page1 - edited');
-      assert.equal(pages[1].slug, '/page1');
-      assert.equal(pages[1].aposMode, 'published');
-      assert.equal(pages[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-
-    it('should update a piece draft and published versions when there is an update key that matches only the existing published doc', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'topic',
-              'title:key': 'topic1',
-              title: 'topic1 - edited',
-              description: 'description1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      const topic = await apos.topic.insert(req, {
-        ...apos.topic.newInstance(),
-        title: 'topic1',
-        description: 'description1',
-        main: '<p><em>rich</em> <strong>text</strong></p>'
-      });
-
-      // check that the draft doc is also updated, even with a different title
-      await apos.doc.db.updateOne(
-        {
-          aposDocId: topic.aposDocId,
-          aposMode: 'draft'
-        },
-        {
-          $set: {
-            title: 'topic1 - draft title that does not match the published title nor the update key'
-          }
-        }
-      );
-
-      await importExportManager.import(req);
-
-      const topics = await apos.doc.db
-        .find({ type: 'topic' })
-        .toArray();
-
-      assert.equal(topics.length, 2);
-
-      assert.equal(topics[0].aposDocId, topic.aposDocId);
-      assert.equal(topics[0].title, 'topic1 - edited');
-      assert.equal(topics[0].slug, 'topic1');
-      assert.equal(topics[0].aposMode, 'draft');
-      assert.equal(topics[0].description, 'description1 - edited');
-      assert.equal(topics[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-      assert.equal(topics[0].modified, false);
-
-      assert.equal(topics[1].aposDocId, topic.aposDocId);
-      assert.equal(topics[1].title, 'topic1 - edited');
-      assert.equal(topics[1].slug, 'topic1');
-      assert.equal(topics[1].aposMode, 'published');
-      assert.equal(topics[1].description, 'description1 - edited');
-      assert.equal(topics[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-
-    it('should update a page draft and published versions when there is an update key that only matches the existing published doc', async function() {
-      csv.input = async () => {
-        return {
-          docs: [
-            {
-              type: 'default-page',
-              'title:key': 'page1',
-              title: 'page1 - edited',
-              main: '<p><em>rich</em> <strong>text</strong> - edited</p>'
-            }
-          ]
-        };
-      };
-
-      const page = await apos.page.insert(req, '_home', 'lastChild', {
-        ...apos.modules['default-page'].newInstance(),
-        title: 'page1',
-        main: '<p><em>rich</em> <strong>text</strong></p>'
-      });
-
-      // check that the draft doc is also updated, even with a different title
-      await apos.doc.db.updateOne(
-        {
-          aposDocId: page.aposDocId,
-          aposMode: 'draft'
-        },
-        {
-          $set: {
-            title: 'page1 - draft title that does not match the published title nor the update key'
-          }
-        }
-      );
-
-      await importExportManager.import(req);
-
-      const pages = await apos.doc.db
-        .find({ type: 'default-page' })
-        .toArray();
-
-      assert.equal(pages.length, 2);
-
-      assert.equal(pages[0].aposDocId, page.aposDocId);
-      assert.equal(pages[0].title, 'page1 - edited');
-      assert.equal(pages[0].slug, '/page1');
-      assert.equal(pages[0].aposMode, 'draft');
-      assert.equal(pages[0].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-      assert.equal(pages[0].modified, false);
-
-      assert.equal(pages[1].aposDocId, page.aposDocId);
-      assert.equal(pages[1].title, 'page1 - edited');
-      assert.equal(pages[1].slug, '/page1');
-      assert.equal(pages[1].aposMode, 'published');
-      assert.equal(pages[1].main.items[0].content, '<p><em>rich</em> <strong>text</strong> - edited</p>');
-    });
-  });
 });
